+++ x/crosschain/keeper/cctx.go
@@ -93,44 +93,6 @@ func (k Keeper) setNonceToCCTX(
 	}
 }
 
-// GetCctxCounter retrieves the current counter value
-func (k Keeper) GetCctxCounter(ctx sdk.Context) uint64 {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), []byte(types.CounterValueKey))
-	storedCounter := store.Get([]byte(types.CounterValueKey))
-
-	return sdk.BigEndianToUint64(storedCounter)
-}
-
-// getNextCctxCounter retrieves and increments the counter for ordering
-func (k Keeper) getNextCctxCounter(ctx sdk.Context) uint64 {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), []byte(types.CounterValueKey))
-	storedCounter := k.GetCctxCounter(ctx)
-
-	nextCounter := storedCounter + 1
-
-	store.Set([]byte(types.CounterValueKey), sdk.Uint64ToBigEndian(nextCounter))
-	return nextCounter
-}
-
-// setCctxCounterIndex sets the CCTX in the counter index
-//
-// note that we use the raw bytes in the index rather than the hex encoded bytes
-// like in the main store
-func (k Keeper) setCctxCounterIndex(ctx sdk.Context, cctx types.CrossChainTx) {
-	counterIndexStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.CounterIndexKey))
-	nextCounter := k.getNextCctxCounter(ctx)
-
-	cctxIndex, err := cctx.GetCCTXIndexBytes()
-	if err != nil {
-		k.Logger(ctx).Error("get cctx index bytes", "err", err)
-		return
-	}
-
-	// must use big endian so most significant bytes are first for sortability
-	nextCounterBytes := sdk.Uint64ToBigEndian(nextCounter)
-	counterIndexStore.Set(nextCounterBytes, cctxIndex[:])
-}
-
 // SetCrossChainTx set a specific cctx in the store from its index
 func (k Keeper) SetCrossChainTx(ctx sdk.Context, cctx types.CrossChainTx) {
 	// only set the updated timestamp if the block height is >0 to allow
