+++ x/crosschain/keeper/grpc_query_cctx.go
@@ -46,7 +46,7 @@ func (k Keeper) CctxAll(c context.Context, req *types.QueryAllCctxRequest) (*typ
 
 	store := ctx.KVStore(k.storeKey)
 	cctxStore := prefix.NewStore(store, types.KeyPrefix(types.CCTXKey))
-	counterStore := prefix.NewStore(store, types.KeyPrefix(types.CounterIndexKey))
+	counterStore := k.getCounterValueStore(ctx)
 
 	if req.Pagination == nil {
 		req.Pagination = &query.PageRequest{}
@@ -57,9 +57,10 @@ func (k Keeper) CctxAll(c context.Context, req *types.QueryAllCctxRequest) (*typ
 
 	var sends []*types.CrossChainTx
 	pageRes, err := query.Paginate(counterStore, req.Pagination, func(_ []byte, value []byte) error {
-		var indexBytes [32]byte
-		copy(indexBytes[:], value[:32])
-		cctxIndex := types.GetCctxIndexFromBytes(indexBytes)
+		cctxIndex, err := types.GetCctxIndexFromArbitraryBytes(value)
+		if err != nil {
+			return err
+		}
 		var cctx types.CrossChainTx
 		cctxBytes := cctxStore.Get(types.KeyPrefix(cctxIndex))
 		if err := k.cdc.Unmarshal(cctxBytes, &cctx); err != nil {
