+++ x/crosschain/types/cctx.go
@@ -269,6 +269,18 @@ func GetCctxIndexFromBytes(sendHash [32]byte) string {
 	return fmt.Sprintf("0x%s", hex.EncodeToString(sendHash[:]))
 }
 
+// GetCctxIndexFromArbitraryBytes converts an arbitrary byte slice to a CCTX index string.
+// Returns an error if the input slice is less than 32 bytes.
+func GetCctxIndexFromArbitraryBytes(sendHash []byte) (string, error) {
+	if len(sendHash) < 32 {
+		return "", fmt.Errorf("input byte slice length %d is less than required 32 bytes", len(sendHash))
+	}
+
+	var indexBytes [32]byte
+	copy(indexBytes[:], sendHash[:32])
+	return GetCctxIndexFromBytes(indexBytes), nil
+}
+
 // NewCCTX creates a new CCTX from a MsgVoteInbound message and a TSS pubkey.
 // It also validates the created cctx
 func NewCCTX(ctx sdk.Context, msg MsgVoteInbound, tssPubkey string) (CrossChainTx, error) {
